Fixed minors including ambiguous calls of the function isnan(), and modified the build scripts

+++ build/linux/make_centos_llvm.sh
@@ -1,12 +1,15 @@
-g++ -c ../../source/CSMP_number_types_main.cpp
-g++ -o CSMP_number_types_main CSMP_number_types_main.o
-./CSMP_number_types_main
-cp ./CSMP_number_types.h ../../source/includes/model/
+#env LM_LICENSE_FILE="../../bin/license.dat"
+#env LD_LIBRARY_PATH="../../bin/data"
 
-bjam -j32 --toolset=clang --build-type=complete --layout=system optimization=speed link=shared threading=multi cxxflags="-std=c++11"
+if [ -f "CSMP_number_type.h" ]
+  then
+    clang++ -c ../../source/CSMP_number_types_main.cpp
+    clang++ -o CSMP_number_types_main CSMP_number_types_main.o
+    ./CSMP_number_types_main
+    cp ./CSMP_number_types.h ../../source/includes/model/
+fi
 
-env LM_LICENSE_FILE="../../bin/license.dat"
-env LD_LIBRARY_PATH="../../bin/data"
+bjam -j2 --toolset=clang --build-type=complete --layout=system optimization=speed link=shared threading=multi cxxflags="-std=c++11"
 
 cp ../../bin/linux/csmptest ../../bin/data/
 cd ../../bin/data
